pyproject.toml
[tool.poetry]
name = "score-simulator-data"
version = "0.0.1"
description = "score-simulator temporary static data storage"
authors = ["tanzhijian <tanzhijianorg@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fusion-stat = "^0.0.10"
python-dotenv = "^1.0.0"
tqdm = "^4.66.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88
